+++ CHANGELOG.md
@@ -22,6 +22,9 @@ For older changes see the [archived Singularity change log](https://github.com/a
 - Apptainer now supports the `riscv64` architecture.
 - Native cgroups v2 resource limits can be specified using the `[unified]` key
   in a cgroups toml file applied via `--apply-cgroups`.
+- The `--no-mount` flag & `APPTAINER_NO_MOUNT` env var can now be used to
+  disable a `bind path` entry from `apptainer.conf` by specifying the
+  absolute path to the destination of the bind.
 
 ### Bug fixes